Use performance.now() over unstable performance profiling for now.
The main reason for this for now is that I'm interested in loading times, rather than really fine grained rendering times. I think render events could be reported through that API later, but another obstacle is that that React code will get stripped out unless it's the timin bundle. Something for later.